fix: clarify composer dock contracts
